Add messagegroup
Messagegroups are a new kind of group. Every user can join and leave such a
group on her own. They are useful to create topics for mails.

E.g. if there is a weekly newsletter we can create a new messagegroup for
it and every user which is interested in receiving the newsletter can join
the corresponding group.