Lock SQLite3 to < 1.4
On my system, I couldn't get `rake db:create` to run on `master`, as the
version of SQLite3 in the Gemfile.lock would result in the following
error:
https://stackoverflow.com/questions/54527277/cant-activate-sqlite3-1-3-6-already-activated-sqlite3-1-4-0

(Ok, not my question, it was 1.4.1 on my machine, but same thing).

Locking SQLite3 to '~> 1.3.6` did fix the issue. It'll probably a great
candidate to upgrade once we're on Rails 5.