Add a script to read all databases with a prefix from the database server
When passing a value via the FOODSOFT_DB_PREFIX environment variable the
FOODSOFT_FOODCOOPS environment variable will be overwritten with a list
of databases starting with the given prefix.

This can be used to dynamically generate the app_config.yml in a multicoop
installation and avoid listing all instances manually.